bench(sirun): model per-hook registration and move validation to setup
Review follow-ups, two P1s:

- The simulated workload only read the returned hooks' count, but every
  production call site registers each returned hook through addHook
  (e.g. azure-cosmos.js). The bench now models that pass - one closure
  and one push per returned hook - because the 144-vs-66 count is a real
  part of what the fix changes. Under this workload the dedupe costs ~8%
  cold (~35 us) on a fully instrumented startup and ~19% warm steady
  state, while registering 78 fewer hooks per startup. The earlier
  'faster warm' reading was an artifact of the count-only model.

- The equivalence assertions ran after the measured window but still
  inside the measured process: sirun resets wall.time/instructions at
  the ready signal and measures until process exit, so the validation's
  26 extra getHooks calls polluted every sample. Validation moved to
  validate.js, wired as a sirun setup command - it runs in its own
  process before the measured process exists, so it can neither warm the
  measured process nor land in the window; a mismatch still aborts the
  run before a single sample is collected (verified with a deliberately
  broken dedupe).

The embedded implementations moved to get-hooks.js, shared verbatim by
index.js and validate.js.
