+++ benchmark/sirun/instrument-get-hooks/README.md
@@ -7,23 +7,36 @@ hook per transform and integrations registered the same (versionRange,
 filePath) several times (144 hooks for the real query set before the fix, 66
 after).
 
-The workload models what production actually does: `getHooks` is called once
-per module name, **lazily**, from integration files that
-`helpers/register.js` only runs when the user's package loads. A traced
-process performs **zero** of these calls at tracer init, and all 13 only when
-every instrumented package family is used. The 13 queries are the complete set
-of call sites in the repo (`ai.js`, `langchain.js`, `mercurius.js`,
-`bullmq.js`, `modelcontextprotocol-sdk.js`, `openai-agents.js`,
-`langgraph.js`, `aws-durable-execution-sdk-js.js`, `azure-cosmos.js`,
-`claude-agent-sdk.js`, and the three in `graphql.js`). There is no hot loop
-to optimize and no fixed startup cost to charge against lookup savings, so
-the bench measures one simulated lazy load pass per iteration.
+The workload models what production actually does, end to end:
+
+- `getHooks` is called once per module name, **lazily**, from integration
+  files that `helpers/register.js` only runs when the user's package loads.
+  A traced process performs **zero** of these calls at tracer init, and all
+  13 only when every instrumented package family is used. The 13 queries
+  are the complete set of call sites in the repo (`ai.js`, `langchain.js`,
+  `mercurius.js`, `bullmq.js`, `modelcontextprotocol-sdk.js`,
+  `openai-agents.js`, `langgraph.js`, `aws-durable-execution-sdk-js.js`,
+  `azure-cosmos.js`, `claude-agent-sdk.js`, and the three in `graphql.js`).
+- Production call sites never stop at the hook count: each one registers
+  every returned hook through `addHook` (for example `azure-cosmos.js`:
+  `for (const hook of getHooks('@azure/cosmos')) addHook(hook, exports =>
+  exports)`). The bench models that pass too — one closure, one push into
+  the per-name list, per returned hook — because the number of returned
+  hooks (144 vs 66) is a real part of what the fix changes.
+
+There is no hot loop to optimize and no fixed startup cost to charge against
+lookup savings, so the bench measures one simulated lazy load pass per
+iteration.
 
 - `scan` — the pre-fix implementation, verbatim: map → filter → map, one
   hook per transform (duplicates included).
 - `deduped` — the fixed implementation, verbatim: one pass, skipping
   transforms whose (versionRange, filePath) was already emitted, with a Set
   for the requested names.
+
+Both live in `get-hooks.js`; `validate.js` runs as a sirun `setup` command
+before the measured process exists, so its equivalence gate can neither warm
+the measured process nor land inside the measured window.
 
 Variants (`meta.json`):
 
@@ -35,22 +48,28 @@ Variants (`meta.json`):
 - `*-warm` — 20 000 simulated startups per process, for steady-state signal
   over the same workload.
 
-Measured (Node v26.2.0, median of fresh processes / warm loops):
+Measured (Node v26.5.0, median of fresh processes / warm loops):
 
-| variant | per startup (13 queries) | hooks resolved |
+| variant | per startup (13 queries + registration) | hooks registered |
 |---|---|---|
-| `scan-cold` | ~175 µs | 144 |
-| `deduped-cold` | ~195 µs | 66 |
-| `scan-warm` | ~34 µs | 144 |
-| `deduped-warm` | ~29 µs | 66 |
+| `scan-cold` | ~435 µs | 144 |
+| `deduped-cold` | ~470 µs | 66 |
+| `scan-warm` | ~22.5 µs | 144 |
+| `deduped-warm` | ~27 µs | 66 |
 
 Zero-lookup startups (the common case) run no `getHooks` code in either
 variant: the fix adds no fixed cost anywhere — nothing is built at require
 time.
 
-Read: the correctness fix costs ~10% cold on a process that loads *all* 13
-integration families (~20 µs, only paid as the packages load), and wins
-~15% warm. A zero-lookup process pays nothing either way.
+Read: the correctness fix costs ~8% (~35 µs) on a *fully instrumented*
+startup, only as the packages load, and ~19% in steady state per startup.
+Warm is the pessimistic lens: production never calls `getHooks` twice for
+the same name in one process, so the cold number is the production cost.
+The warm gap is real and instructive — the dedupe pays a Set lookup per
+list entry plus a (versionRange, filePath) key string per match, which
+costs more after JIT than the scan's extra duplicate allocations, even
+while saving 78 registrations per startup. A process using a single
+integration family pays one query's share of the delta: a few µs, once.
 
 Run with:
