+++ benchmark/sirun/instrument-get-hooks/meta.json
@@ -1,5 +1,6 @@
 {
   "name": "instrument-get-hooks",
+  "setup": "node validate.js",
   "run": "node index.js",
   "run_with_affinity": "bash -c \"taskset -c $CPU_AFFINITY node index.js\"",
   "cachegrind": false,
